Move subcommand folders to be in the same folder as parent command
