+++ src/commands/Configuration/configure/automod/settings.ts
@@ -1,11 +1,11 @@
 import { Message, MessageEmbed } from "discord.js";
-import DatabaseManager from "../../../utils/managers/DatabaseManager";
-import Command from "../../../structures/Command";
-import SubCommand from "../../../structures/SubCommand";
-import AutoModConfiguration from "../../../utils/types/AutoModConfiguration";
-import { embedColor } from "../../../Config";
-import Emotes from "../../../emotes.json";
-import BulbBotClient from "../../../structures/BulbBotClient";
+import DatabaseManager from "../../../../utils/managers/DatabaseManager";
+import Command from "../../../../structures/Command";
+import SubCommand from "../../../../structures/SubCommand";
+import AutoModConfiguration from "../../../../utils/types/AutoModConfiguration";
+import { embedColor } from "../../../../Config";
+import Emotes from "../../../../emotes.json";
+import BulbBotClient from "../../../../structures/BulbBotClient";
 
 const databaseManager: DatabaseManager = new DatabaseManager();
