+++ src/commands/Configuration/configure/override.ts
@@ -2,12 +2,12 @@ import Command from "../../../structures/Command";
 import SubCommand from "../../../structures/SubCommand";
 import BulbBotClient from "../../../structures/BulbBotClient";
 
-import create from "../override/create";
-import disable from "../override/disable";
-import edit from "../override/edit";
-import enable from "../override/enable";
-import list from "../override/list";
-import remove from "../override/remove";
+import create from "./override/create";
+import disable from "./override/disable";
+import edit from "./override/edit";
+import enable from "./override/enable";
+import list from "./override/list";
+import remove from "./override/remove";
 
 export default class extends SubCommand {
 	constructor(client: BulbBotClient, parent: Command) {
