+++ src/commands/Configuration/configure/override/enable.ts
@@ -1,15 +1,15 @@
 import { Message, Snowflake } from "discord.js";
-import ClearanceManager from "../../../utils/managers/ClearanceManager";
-import Command from "../../../structures/Command";
-import SubCommand from "../../../structures/SubCommand";
-import BulbBotClient from "../../../structures/BulbBotClient";
+import ClearanceManager from "../../../../utils/managers/ClearanceManager";
+import Command from "../../../../structures/Command";
+import SubCommand from "../../../../structures/SubCommand";
+import BulbBotClient from "../../../../structures/BulbBotClient";
 
 const clearanceManager: ClearanceManager = new ClearanceManager();
 
 export default class extends SubCommand {
 	constructor(client: BulbBotClient, parent: Command) {
 		super(client, parent, {
-			name: "disable",
+			name: "enable",
 			minArgs: 1,
 			maxArgs: 1,
 			argList: ["command:string"],
@@ -18,24 +18,24 @@ export default class extends SubCommand {
 	}
 
 	async run(message: Message, args: string[]): Promise<void | Message> {
-		const command: string = args[0];
+		const command = args[0];
 		const cTemp = this.client.commands.get(command.toLowerCase()) || this.client.commands.get(<string>this.client.aliases.get(command.toLowerCase()));
-		if (cTemp === undefined || cTemp.name === undefined)
+		if (!cTemp || cTemp.name === undefined)
 			return message.channel.send(
 				await this.client.bulbutils.translate("global_not_found", message.guild?.id, {
 					type: await this.client.bulbutils.translate("global_not_found_types.cmd", message.guild?.id, {}),
-					arg_provided: args[0],
 					arg_expected: "command:string",
+					arg_provided: args[0],
 					usage: this.usage,
 				}),
 			);
 
 		if ((await clearanceManager.getCommandOverride(<Snowflake>message.guild?.id, cTemp.name)) !== undefined) {
-			await clearanceManager.setEnabled(<Snowflake>message.guild?.id, cTemp.name, false);
+			await clearanceManager.setEnabled(<Snowflake>message.guild?.id, cTemp.name, true);
 		} else {
-			await clearanceManager.createCommandOverride(<Snowflake>message.guild?.id, cTemp.name, false, cTemp.clearance);
+			return message.channel.send(await this.client.bulbutils.translate("override_nonexistent_command", message.guild?.id, { command }));
 		}
 
-		await message.channel.send(await this.client.bulbutils.translate("override_disable_success", message.guild?.id, { command }));
+		await message.channel.send(await this.client.bulbutils.translate("override_enable_success", message.guild?.id, { command }));
 	}
 }
