+++ src/commands/Configuration/configure/override/list.ts
@@ -1,10 +1,10 @@
-import * as Emotes from "../../../emotes.json";
-import { embedColor } from "../../../Config";
+import * as Emotes from "../../../../emotes.json";
+import { embedColor } from "../../../../Config";
 import { Message, MessageEmbed, Snowflake } from "discord.js";
-import ClearanceManager from "../../../utils/managers/ClearanceManager";
-import Command from "../../../structures/Command";
-import SubCommand from "../../../structures/SubCommand";
-import BulbBotClient from "../../../structures/BulbBotClient";
+import ClearanceManager from "../../../../utils/managers/ClearanceManager";
+import Command from "../../../../structures/Command";
+import SubCommand from "../../../../structures/SubCommand";
+import BulbBotClient from "../../../../structures/BulbBotClient";
 
 const clearanceManager: ClearanceManager = new ClearanceManager();
